requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot==20.7
python-dotenv==1.0.0
selenium==4.16.0
webdriver_manager==4.0.1 